feat: stm32 pwm and system counter configured
- fixed avr pwm disable bug where adc does not work

+++ src/hal_stm32/stm32_hal.h
@@ -5,7 +5,7 @@
 
 #include "stm32f1xx_hal.h"
 
-uint16_t getTimerVal(void);
+uint32_t getTimerVal(void);
 
 uint32_t disableInterrupts(void);
